+++ src/effects/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod color;
 pub(crate) mod complex;
 pub(crate) mod dynamic;
 pub(crate) mod fire;
+pub(crate) mod motion;
 pub(crate) mod rainbow;
 pub(crate) mod scanner;
 pub(crate) mod twinkle;